Update preprocessing of data to treat NULLs as zeros, and remove Alaska
Emailed the USDA in regards to the numerous NULLs present in the dataset. Their response indicated the NULLs should be treated as zeros, and that they did not indicate missing or unavailable data.
Temporarily removing Alaska from the dataset, as they have recently changed some counties. The census tracts will need to be mapped as was done with Connecticut before including them back in the dataset.